[Refactor] move widget to BikeApp directory
